+++ src/main/java/com/capstone/liveAloneComunity/service/auth/AuthService.java
@@ -6,7 +6,7 @@ import com.capstone.liveAloneComunity.entity.member.Member;
 import com.capstone.liveAloneComunity.entity.RefreshToken;
 import com.capstone.liveAloneComunity.exception.authentication.LogInAgainException;
 import com.capstone.liveAloneComunity.repository.member.MemberRepository;
-import com.capstone.liveAloneComunity.repository.RefreshTokenRepository;
+import com.capstone.liveAloneComunity.repository.refreshToken.RefreshTokenRepository;
 import com.capstone.liveAloneComunity.service.member.MemberValidator;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
